pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "talenWF"
version = "0.1.0"
description = "Modern TALEN window finder compatible with MitoEdit. Made as a part of the St. Jude KIDS25 BioHackathon Project. This tool is an effort to revamp and modernize the TAL Effector-Nucleotide Targeter (TALE-NT) 2.0 (Doyle et al., 2012, Nucleic Acids Research)"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Ishaan Gupta", email = "i3gupta@ucsd.edu" }
]
maintainers = [
    { name = "Ishaan Gupta", email = "i3gupta@ucsd.edu" }
]
license = { text = "MIT" }
keywords = ["TALEN", "genome editing", "bioinformatics", "mitochondria"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
  "biopython>=1.80",
  "pandas>=2.0.0",
  "numpy>=1.20.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=6.0",
  "pytest-cov>=2.0",
  "black>=22.0",
  "flake8>=4.0",
]

[project.urls]
Homepage = "https://github.com/IshaanSD/talenWF"
"Bug Reports" = "https://github.com/IshaanSD/talenWF/issues"
"Source" = "https://github.com/IshaanSD/talenWF"
"Documentation" = "https://github.com/IshaanSD/talenWF#readme"
"Original TALE-NT 2.0 Paper" = "https://doi.org/10.1093/nar/gks608"

[project.scripts]
talenWF-findtal = "talenWF.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["talenWF*"]


